Development

+++ app/Livewire/BeritaTable.php
@@ -6,6 +6,8 @@ use App\Models\Berita;
 use Barryvdh\DomPDF\Facade\Pdf;
 use Illuminate\Support\Carbon;
 use Illuminate\Database\Eloquent\Builder;
+use Illuminate\Support\Facades\Storage;
+use Masmerise\Toaster\Toastable;
 use PowerComponents\LivewirePowerGrid\Button;
 use PowerComponents\LivewirePowerGrid\Column;
 use PowerComponents\LivewirePowerGrid\Detail;
@@ -21,15 +23,13 @@ use PowerComponents\LivewirePowerGrid\Traits\WithExport;
 final class BeritaTable extends PowerGridComponent
 {
     use WithExport;
+    use Toastable;
 
     public function setUp(): array
     {
         $this->showCheckBox();
 
-        return [
-            Exportable::make('export')
-                ->striped()
-                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV),
+        $setUp = [
             Header::make()->showSearchInput(),
             Footer::make()
                 ->showPerPage()
@@ -38,20 +38,25 @@ final class BeritaTable extends PowerGridComponent
                 ->showCollapseIcon()
                 ->view('details.berita-detail'),
         ];
+
+        if (auth()->user()->can('export')) {
+            $setUp[] = Exportable::make('export')
+                ->striped()
+                ->type(Exportable::TYPE_XLS, Exportable::TYPE_CSV);
+        }
+
+        return $setUp;
     }
 
     public function datasource(): Builder
     {
-        return Berita::query()
-            ->leftJoin('users', 'berita.user_id', '=', 'users.id')
-            ->select('berita.*', 'users.name as name','users.no_hp as no_hp','users.alamat as alamat');
-        ;
+        return Berita::query()->with('user');
     }
 
     public function relationSearch(): array
     {
         return [
-            'users'=> ['name','no_hp','alamat']
+            'user' => ['name']
         ];
     }
 
@@ -59,37 +64,29 @@ final class BeritaTable extends PowerGridComponent
     {
         return PowerGrid::fields()
             ->add('id')
-            ->add('name')
-            ->add('judul_berita')
-            ->add('deskripsi_berita')
+            ->add('penulis', fn ($row) => $row->user->name)
+            ->add('judul')
+            ->add('deskripsi')
             ->add('foto')
-            ->add('no_berita');
-         
+            ->add('no_berita')
+            ->add('created_at_formatted', fn ($row) => $row->created_at->format('d-m-Y'));
     }
     public function columns(): array
     {
         return [
-            Column::make('Id','id')
+            Column::make('Id', 'id')
                 ->searchable()
                 ->sortable(),
 
-            Column::make('Nama ','name')
-                ->searchable()
-                ->sortable(),
-                
-            Column::make('Judul Berita','judul_berita')
+            Column::make('Nama ', 'penulis')
                 ->searchable()
                 ->sortable(),
 
-            // Column::make('Deskripsi Berita','deskripsi_berita')
-            //     ->searchable()
-            //     ->sortable(),
+            Column::make('Judul Berita', 'judul')
+                ->searchable()
+                ->sortable(),
 
-            // Column::make('Foto','foto')
-            //     ->searchable()
-            //     ->sortable(),
-
-            Column::make('No Berita','no_berita')
+            Column::make('Created At', 'created_at_formatted')
                 ->searchable()
                 ->sortable(),
 
@@ -99,59 +96,65 @@ final class BeritaTable extends PowerGridComponent
 
     public function filters(): array
     {
-        return [
-        ];
-    }
-
-    #[\Livewire\Attributes\On('edit')]
-    public function edit($rowId): void
-    {
-        $this->js('alert('.$rowId.')');
+        return [];
     }
 
     public function actions(\App\Models\Berita $row): array
     {
-        return [
-            Button::add('edit')
+        $actions = [];
+
+        if (auth()->user()->can('update')) {
+            $actions[] =
+                Button::add('edit')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m16.862 4.487 1.687-1.688a1.875 1.875 0 1 1 2.652 2.652L10.582 16.07a4.5 4.5 0 0 1-1.897 1.13L6 18l.8-2.685a4.5 4.5 0 0 1 1.13-1.897l8.932-8.931Zm0 0L19.5 7.125M18 14v4.75A2.25 2.25 0 0 1 15.75 21H5.25A2.25 2.25 0 0 1 3 18.75V8.25A2.25 2.25 0 0 1 5.25 6H10" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->openModal('berita-form', ['rowId' => $row->id]),
-            Button::add('delete')
+                ->openModal('berita-form', ['rowId' => $row->id]);
+        }
+
+        if (auth()->user()->can('delete')) {
+            $actions[] = Button::add('delete')
                 ->slot('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="m14.74 9-.346 9m-4.788 0L9.26 9m9.968-3.21c.342.052.682.107 1.022.166m-1.022-.165L18.16 19.673a2.25 2.25 0 0 1-2.244 2.077H8.084a2.25 2.25 0 0 1-2.244-2.077L4.772 5.79m14.456 0a48.108 48.108 0 0 0-3.478-.397m-12 .562c.34-.059.68-.114 1.022-.165m0 0a48.11 48.11 0 0 1 3.478-.397m7.5 0v-.916c0-1.18-.91-2.164-2.09-2.201a51.964 51.964 0 0 0-3.32 0c-1.18.037-2.09 1.022-2.09 2.201v.916m7.5 0a48.667 48.667 0 0 0-7.5 0" />
                 </svg>
                 ')
                 ->id()
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700')
-                ->dispatch('delete', ['rowId' => $row->id])
-        ];
+                ->dispatch('delete', ['rowId' => $row->id]);
+        }
+        return $actions;
     }
 
     public function header(): array
     {
-        return [
-            Button::add('add-berita')
+        $header = [];
+
+        if (auth()->user()->can('create')) {
+            $header[] = Button::add('add-berita')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M12 4.5v15m7.5-7.5h-15" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->openModal('berita-form', []),
-            Button::add('export-pdf')
+                ->openModal('berita-form', []);
+        }
+
+        if (auth()->user()->can('export')) {
+            $header[] = Button::add('export-pdf')
                 ->slot(__('<svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" class="w-5 h-5">
                 <path stroke-linecap="round" stroke-linejoin="round" d="M6.72 13.829c-.24.03-.48.062-.72.096m.72-.096a42.415 42.415 0 0 1 10.56 0m-10.56 0L6.34 18m10.94-4.171c.24.03.48.062.72.096m-.72-.096L17.66 18m0 0 .229 2.523a1.125 1.125 0 0 1-1.12 1.227H7.231c-.662 0-1.18-.568-1.12-1.227L6.34 18m11.318 0h1.091A2.25 2.25 0 0 0 21 15.75V9.456c0-1.081-.768-2.015-1.837-2.175a48.055 48.055 0 0 0-1.913-.247M6.34 18H5.25A2.25 2.25 0 0 1 3 15.75V9.456c0-1.081.768-2.015 1.837-2.175a48.041 48.041 0 0 1 1.913-.247m10.5 0a48.536 48.536 0 0 0-10.5 0m10.5 0V3.375c0-.621-.504-1.125-1.125-1.125h-8.25c-.621 0-1.125.504-1.125 1.125v3.659M18 10.5h.008v.008H18V10.5Zm-3 0h.008v.008H15V10.5Z" />
                 </svg>
                 '))
                 ->class('pg-btn-white dark:ring-pg-primary-600 dark:border-pg-primary-600 dark:hover:bg-pg-primary-700 dark:ring-offset-pg-primary-800 dark:text-pg-primary-300 dark:bg-pg-primary-700 w-full')
-                ->dispatch('exportPdf', []),
-        ];
+                ->dispatch('exportPdf', []);
+        }
+        return $header;
     }
 
-    public function gitlisteners()
+    public function getlisteners()
     {
         return array_merge(
             parent::getListeners(),
@@ -159,48 +162,40 @@ final class BeritaTable extends PowerGridComponent
                 'delete',
                 'exportPdf',
                 'edit',
-                'berita-updated' => '$refresh',
-            ] 
+                'beritaUpdated' => '$refresh',
+            ]
         );
     }
 
-     // Function to export PDF using DomPDF
-     public function exportPdf()
-     {
-         $path = public_path() . '/pdf';
-         // Mendapatkan datasource
-         $datasource = $this->datasource()->get();
-         // Membuat folder pdf jika belum ada
-         if (!file_exists($path)) {
-             mkdir($path, 0777, true);
-         }
-         // Membuat file pdf
-         $pdf = Pdf::loadView('pdf.berita', ['datasource' => $datasource]);
-         // Menyimpan file pdf ke folder pdf
-         $pdf->save($path . '/berita.pdf');
-         // Menampilkan file pdf
-         return response()->download($path . '/berita.pdf');
-     }
- 
-     // Function to delete data
-     public function delete($rowId)
-     {
-         $berita = Berita::findOrFail($rowId);
-         // Detach all associated users
-         $berita->user()->detach();
-         $berita->delete();
-     }
-
-
-    /*
-    public function actionRules($row): array
+    // Function to export PDF using DomPDF
+    public function exportPdf()
     {
-       return [
-            // Hide button edit for ID 1
-            Rule::button('edit')
-                ->when(fn($row) => $row->id === 1)
-                ->hide(),
-        ];
+        $path = public_path() . '/pdf';
+        // Mendapatkan datasource
+        $datasource = $this->datasource()->get();
+        // Membuat folder pdf jika belum ada
+        if (!file_exists($path)) {
+            mkdir($path, 0777, true);
+        }
+        // Membuat file pdf
+        $pdf = Pdf::loadView('pdf.berita', ['datasource' => $datasource]);
+        // Menyimpan file pdf ke folder pdf
+        $pdf->save($path . '/berita.pdf');
+        // Menampilkan file pdf
+        return response()->download($path . '/berita.pdf');
     }
-    */
+
+    // Function to delete data
+    public function delete($rowId)
+    {
+        $berita = Berita::findOrFail($rowId);
+
+        if ($berita->foto) {
+            Storage::disk('public')->delete('berita/' . $berita->foto);
+        }
+
+        $berita->delete();
+
+        $this->success('Berita berhasil dihapus');
+    }
 }
